+++ Changelog.md
@@ -5,7 +5,8 @@
  - Added `~=` operator for `Regex`
  - Added the `!!` operator for calling a never-returning function when optional unwrapping fails
  - Added `AnyType` and `Nothing` typealiases for `Any.Type` and `Void`, respectively
- - Added `sum()` and `product()` methods
+ - Added `sum()` and `product()` methods to `Numeric` sequences
+ - Added `compacted()` method to `Sequence`
 
 ## 2.3.0